+++ HTMLKit/HTMLInputStreamReader.m
@@ -82,7 +82,7 @@
 		return LINE_FEED;
 	}
 	if (CFStringIsSurrogateLowCharacter(nextInputCharacter)) {
-		NSString *reason = [NSString stringWithFormat:@"Non-Unicode character found (an isolated low surrogate: 0x%X)", nextInputCharacter];
+		NSString *reason = [NSString stringWithFormat:@"Non-Unicode character found (an isolated low surrogate: 0x%X)", (unsigned int)nextInputCharacter];
 		[self emitParseError:reason];
 		return nextInputCharacter;
 	}
@@ -90,7 +90,7 @@
 	if (CFStringIsSurrogateHighCharacter(nextInputCharacter)) {
 		UniChar surrogateLow = CFStringGetCharacterFromInlineBuffer(&_buffer, _location + 1);
 		if (CFStringIsSurrogateLowCharacter(surrogateLow) == NO) {
-			NSString *reason = [NSString stringWithFormat:@"Non-Unicode character found (an isolated high surrogate: 0x%X)", nextInputCharacter];
+			NSString *reason = [NSString stringWithFormat:@"Non-Unicode character found (an isolated high surrogate: 0x%X)", (unsigned int)nextInputCharacter];
 			[self emitParseError:reason];
 			return nextInputCharacter;
 		}
@@ -100,7 +100,7 @@
 	}
 
 	if (isControlOrUndefinedCharacter(nextInputCharacter)) {
-		NSString *reason = [NSString stringWithFormat:@"A control/undefined character found: (0x%X)", nextInputCharacter];
+		NSString *reason = [NSString stringWithFormat:@"A control/undefined character found: (0x%X)", (unsigned int)nextInputCharacter];
 		[self emitParseError:reason];
 	}
